+++ src/main/java/com/javainuse/springbootsecurity/repository/EmployeeRepo.java
@@ -5,6 +5,6 @@ import com.javainuse.springbootsecurity.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface EmployeeRepo extends JpaRepository<Employee, Integer> {
-    Employee findById(int id);
+public interface EmployeeRepo extends JpaRepository<Employee, Long> {
+    Employee findById(long id);
 }
